install translated manpages
now if you are wondering why I add that to .install, well... we always have
them in install. except in this package, where the manpage is actually
in manpages. I have no clue why we usually put them in install. I also
have no clue why it is in .manpages in this package.
The day any of this starts to make sense drinks will be on the house...